exracted data to js files
