+++ data/seeds/006_add_resources_table.js
@@ -1,79 +1,4 @@
+const { resources } = require("../utils")
 exports.seed = function(knex) {
-    return knex("resources").insert([
-        {
-            type: "google_doc",
-            title: "Action verbs for technical resumes",
-            tasks_id: 1,
-            url:
-                "https://docs.google.com/document/d/1wZkDPBWtQZDGGdvStD61iRx_jOWVlIyyQl9UOYHtZgA/edit"
-        },
-        {
-            type: "unspecified",
-            title: "Power statements article",
-            tasks_id: 1,
-            url:
-                "https://www.linkedin.com/pulse/20140929001534-24454816-my-personal-formula-for-a-better-resume/"
-        },
-        {
-            type: "google_doc",
-            title: "'Lambda is…' paragraphs",
-            tasks_id: 1,
-            url:
-                "https://docs.google.com/document/d/19OxIgJYkLMq4c1o5zHu1Na4a3PYcyutOosVfg6a03RI/edit"
-        },
-        {
-            type: "google_doc",
-            title: "Resume Deep Dive",
-            tasks_id: 1,
-            url:
-                "https://docs.google.com/document/d/1BOwmlkiPJH_96dnjI3nAW8ok-BOR8hkl5Us70bwf24U/edit"
-        },
-        {
-            type: "unspecified",
-            title: "Resume  Rubric",
-            tasks_id: 1,
-            url:
-                "https://drive.google.com/file/d/1h0CCENtwDuNxVwqIG8-P6O2sZj8S4BPg/view"
-        },
-        {
-            type: "google_doc",
-            title: "Resume tips for UX Designers",
-            tasks_id: 1,
-            url:
-                "https://docs.google.com/document/d/1yHAw9CbXjmA5TJwI8j84uS-w5tojC8y7eeoGmmoYzMc/edit"
-        },
-        {
-            type: "google_doc",
-            title: "Resume tips for Mobile Developers",
-            tasks_id: 1,
-            url:
-                "https://docs.google.com/document/d/1bvqn0aFGpvkaKVfmBDAhtmoWlAlDoRTD5DuiMXwsJlQ/edit"
-        },
-        {
-            type: "google_doc",
-            title: "Resume tips for Web Developers",
-            tasks_id: 1,
-            url:
-                "https://docs.google.com/document/d/1RHpYdEqrvSPL3Z4dIHLVaYyju06aGrApBI6RMyAqFg0/edit"
-        },
-        {
-            type: "google_doc",
-            title: "Resume tips for Data Scientists",
-            tasks_id: 1,
-            url:
-                "https://docs.google.com/document/d/1bYAuFyZfa-BIAYUf529Px3ULNOKZwAvUFRxDqlOCvaE/edit"
-        },
-        {
-            type: "youtube_vid",
-            title: "Link to Recorded Video",
-            tasks_id: 3,
-            url: "https://youtu.be/osyavDyNeQ4"
-        },
-        {
-            type: "youtube_vid",
-            title: "Link to Recorded Video",
-            tasks_id: 4,
-            url: "https://youtu.be/AdAA9d5vCaE"
-        }
-    ])
+    return knex("resources").insert(resources)
 }
