+++ data/utils/tracks.js
@@ -1,0 +1,9 @@
+module.exports = {
+    tracks: [
+        { title: "Full-Stack Web" },
+        { title: "iOS" },
+        { title: "Data Science" },
+        { title: "Android" },
+        { title: "UX Design" }
+    ]
+}
