+++ data/utils/user_steps_completed.js
@@ -1,0 +1,7 @@
+module.exports = {
+    userStepsCompleted: [
+        { user_id: 1, steps_id: 1 },
+        { user_id: 1, steps_id: 2 },
+        { user_id: 1, steps_id: 6 }
+    ]
+}
